Consolidated new lwcLogger component with the original logger component
